+++ src/pages/DataExporter/TimeTracking/index.vue
@@ -136,7 +136,7 @@
 </template>
 
 <script setup lang="ts">
-import { computed, ref } from 'vue';
+import { computed, CSSProperties, ref, watch } from 'vue';
 import { getAppKey } from '../../../components/settings';
 import {
   clearToken,
@@ -159,6 +159,11 @@ import { ExportToCsv } from 'export-to-csv';
 import UILoader from '../../../components/UILoader.vue';
 import { getSubscriptionStatus } from '../../../components/optro';
 import UIOptroStatus from '../../../components/UIOptroStatus.vue';
+import { setStorage, getStorage } from '../../../utils/local-storage';
+
+interface Settings {
+  columns: string[];
+}
 
 const isAuthorized = ref(false);
 const memberOptions = ref<Option[]>();
@@ -207,8 +212,23 @@ const listById: {
   [key: string]: Trello.PowerUp.List;
 } = {};
 
-// TODO: Find CSS.Properties types package
-const columnStyle: { [key: keyof ApiCardRowData]: any } = {
+const settings = computed<Settings>(() => {
+  return {
+    columns: columns.value
+  };
+});
+
+watch(settings, () => {
+  setStorage('export-time-tracking', settings.value);
+});
+
+const currentSettings = getStorage<Settings>('export-time-tracking');
+
+if (currentSettings?.columns) {
+  columns.value = currentSettings.columns;
+}
+
+const columnStyle: { [key: keyof ApiCardRowData]: CSSProperties } = {
   'card.description': {
     maxWidth: '200px',
     overflow: 'hidden',
@@ -621,7 +641,9 @@ async function getData() {
     ).then<Trello.PowerUp.Card[]>((res) => res.json());
 
     const boardData = await fetch(
-      `https://api.trello.com/1/boards/${board.id}?fields=name&key=${getAppKey()}&token=${token}&r=${new Date().getTime()}`
+      `https://api.trello.com/1/boards/${
+        board.id
+      }?fields=name&key=${getAppKey()}&token=${token}&r=${new Date().getTime()}`
     ).then<Trello.PowerUp.Board>((res) => res.json());
 
     cards = data.map<ApiCard>((card) => {